move stories to realtive to compoent
